remove not needed readme and selections
