+++ package.json
@@ -26,7 +26,6 @@
     "react-full-screen": "^1.1.1",
     "react-hotkeys": "^2.0.0",
     "react-i18next": "^14.0.5",
-    "react-markdown": "^9.0.0",
     "react-remove-scroll": "^2.5.7",
     "react-select": "^5.7.7",
     "react-use": "^17.4.0",
