+++ src/DemoSite/styles.js
@@ -5,9 +5,9 @@ export default {
     backgroundColor: "#f8f8f8",
     display: "flex",
     alignItems: "center",
-    "& .button": { margin: 5 },
+    "& .button": {margin: 5},
   },
-  select: { width: 240, fontSize: 14 },
+  select: {width: 240, fontSize: 14},
   contentArea: {
     padding: 10,
   },
