+++ src/index.jsx
@@ -3,7 +3,6 @@ import React from "react"
 import {createRoot} from "react-dom/client"
 import Theme from "./Theme"
 import DemoSite from "./DemoSite"
-import LandingPage from "./LandingPage"
 import "./site.css"
 
 
@@ -12,7 +11,7 @@ const Site = () => {
     .replace(/\/$/, "")
     .split("/")
     .slice(-1)[0]
-  return <Theme>{path === "demo" ? <DemoSite /> : <LandingPage />}</Theme>
+  return <Theme><DemoSite /></Theme>
 }
 
 const container = document.getElementById("app")
